backend/pyproject.toml
[project]
name = "agentic-data-quality-platform"
version = "0.1.0"
description = "AI-powered data quality platform with Agno agents"
requires-python = ">=3.11"
dependencies = [
    # Core FastAPI and Web Framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    # Database and ORM
    "sqlalchemy>=2.0.23",
    # Azure OpenAI
    "openai>=1.3.7",
    # Data Processing
    "pandas>=2.1.4",
    "numpy>=1.25.2",
    # Job Scheduling and Background Tasks
    "python-multipart>=0.0.6",
    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "agno>=1.7.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
